requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
pymysql
alembic
pydantic-settings
python-multipart
httpx
beautifulsoup4
playwright
apscheduler
python-docx
portalocker
pywin32; platform_system == "Windows"
cryptography
pyyaml
pytest
pytest-asyncio
pytest-playwright
respx
dingtalk-stream
